requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
sentence-transformers
torch
requests